LPS-133858 Use getWidgetsURL instead of raw widgets object
