+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/actions/types.js
@@ -25,7 +25,6 @@ export const CHANGE_MASTER_LAYOUT = 'CHANGE_MASTER_LAYOUT';
 export const DELETE_FRAGMENT_ENTRY_LINK_COMMENT =
 	'DELETE_FRAGMENT_ENTRY_LINK_COMMENT';
 export const DELETE_ITEM = 'DELETE_ITEM';
-export const DELETE_WIDGETS = 'DELETE_WIDGETS';
 export const DUPLICATE_ITEM = 'DUPLICATE_ITEM';
 export const EDIT_FRAGMENT_ENTRY_LINK_COMMENT =
 	'EDIT_FRAGMENT_ENTRY_LINK_COMMENT';
