+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/services/WidgetService.js
@@ -51,4 +51,8 @@ export default {
 			{requestGenerateDraft: true}
 		);
 	},
+
+	getWidgets() {
+		return serviceFetch(config.getWidgetsURL, {}, () => {});
+	},
 };
